+++ lib/sidekiq/worker/job/server_middleware.rb
@@ -6,7 +6,15 @@ module Sidekiq
         # @param job    [Hash]   the full job payload (https://github.com/mperham/sidekiq/wiki/Job-Format).
         # @param _queue [String] the name of the queue the job was pulled from.
         def call(worker, job, _queue)
-          worker.job = Sidekiq::Worker::Job.new(job)
+          worker.job = Sidekiq::Worker::Job.new(
+            worker_name: job.fetch('class'),
+            queue_name:  job.fetch('queue'),
+            id:          job.fetch('jid'),
+            args:        job.fetch('args'),
+            created_at:  job.fetch('created_at'),
+            enqueued_at: job.fetch('enqueued_at')
+          )
+
           yield
         end
       end