+++ spec/sidekiq/worker/job/server_middleware_spec.rb
@@ -2,18 +2,15 @@ describe Sidekiq::Worker::Job::ServerMiddleware do
   let(:middleware) { described_class.new }
 
   describe '#call' do
-    let(:worker_class) do
-      Class.new do
-        include Sidekiq::Worker
-      end
+    let(:worker) do
+      Class.new { include Sidekiq::Worker }.new
     end
 
-    let(:worker) { worker_class.new }
-    let(:queue)  { 'queue_name' }
+    let(:queue) { 'queue' }
     let(:job) do
       {
         'class'       => 'SomeWorker',
-        'queue'       => 'default',
+        'queue'       => queue,
         'jid'         => 'b4a577edbccf1d805744efa9',
         'args'        => [1, 'arg', true],
         'created_at'  => 1234567890,
@@ -23,37 +20,21 @@ describe Sidekiq::Worker::Job::ServerMiddleware do
 
     it 'sets the job parameter' do
       middleware.call(worker, job, queue) {}
+
       expect(worker.job).to be_instance_of(Sidekiq::Worker::Job)
     end
 
-    it 'sets attribute worker_name for the job' do
+    it 'sets attributes to the job' do
       middleware.call(worker, job, queue) {}
-      expect(worker.job.worker_name).to eq('SomeWorker')
-    end
 
-    it 'sets attribute queue_name for the job' do
-      middleware.call(worker, job, queue) {}
-      expect(worker.job.queue_name).to eq('default')
-    end
-
-    it 'sets attribute id for the job' do
-      middleware.call(worker, job, queue) {}
-      expect(worker.job.id).to eq('b4a577edbccf1d805744efa9')
-    end
-
-    it 'sets attribute args for the job' do
-      middleware.call(worker, job, queue) {}
-      expect(worker.job.args).to eq([1, 'arg', true])
-    end
-
-    it 'sets attribute created_at for the job' do
-      middleware.call(worker, job, queue) {}
-      expect(worker.job.created_at).to eq(1234567890)
-    end
-
-    it 'sets attribute enqueued_at for the job' do
-      middleware.call(worker, job, queue) {}
-      expect(worker.job.enqueued_at).to eq(1234567891)
+      expect(worker.job).to have_attributes(
+        worker_name: 'SomeWorker',
+        queue_name:  'queue',
+        id:          'b4a577edbccf1d805744efa9',
+        args:        [1, 'arg', true],
+        created_at:  1234567890,
+        enqueued_at: 1234567891
+      )
     end
 
     context 'when parameter class is missing' do