+++ spec/sidekiq/worker_spec.rb
@@ -1,12 +1,8 @@
 describe Sidekiq::Worker do
   describe '#job' do
-    let(:worker_class) do
-      Class.new do
-        include Sidekiq::Worker
-      end
+    let(:worker) do
+      Class.new { include Sidekiq::Worker }.new
     end
-
-    let(:worker) { worker_class.new }
 
     it 'returns job attribute' do
       expect(worker.job).to eq(nil)