+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.1.0 (2020-08-10)
+
+  * Change initialize.
+  * Downgrade the required sidekiq version.
+  * Add a method to get the sidekiq queue.
+
 ## 0.0.1 (2020-08-06)
 
   * Gem initialization.